simplified feedback and game over
