--- src/main/java/ch/heigvd/dai/client/Miss.java
@@ -1,17 +1,0 @@
-package ch.heigvd.dai.client;
-
-public class Miss extends Instruction {
-  public Miss() {
-    super("MISS");
-  }
-
-  @Override
-  protected String execute(String[] arguments) {
-    System.out.println("MISS");
-    enemyBoard.getCell(x, y).hit();
-
-    System.out.println("Enemy board\n" + enemyBoard);
-
-    return null;
-  }
-}
