--- src/main/java/ch/heigvd/dai/client/Win.java
@@ -1,15 +1,0 @@
-package ch.heigvd.dai.client;
-
-public class Win extends Instruction {
-
-  public Win() {
-    super("WIN");
-  }
-
-  @Override
-  protected String execute(String[] arguments) {
-    previous_instruction = this;
-    System.out.println("You won.");
-    return null;
-  }
-}
